requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
lxml==5.3.2
pyyaml==6.0.2
jsonpath-ng==1.7.0
pydantic==2.11.1
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
